@@ -1,9 +1,12 @@
 cmake_minimum_required(VERSION 3.20)
 project(fvm LANGUAGES CXX)
 
-set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD 23)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
+
+set(CMAKE_BUILD_RPATH "/usr/local/gcc-14.1.0/lib64")
+set(CMAKE_INSTALL_RPATH "/usr/local/gcc-14.1.0/lib64")
 
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
